Work in progress on average goals by season
