+++ old_stat_tracker.rb
@@ -69,70 +69,70 @@ class StatTracker
   #   result
   # end
 
-  def average_goals_per_game
-    (total_goals.to_f / total_number_of_games).round(2)
-  end
+  # def average_goals_per_game
+  #   (total_goals.to_f / total_number_of_games).round(2)
+  # end
+  #
+  # def total_number_of_goals
+  #   home_goals.count + away_goals.count
+  # end
+  #
+  # def home_goals
+  #   games.find_all do |game|
+  #     game['away_goals'].to_i
+  #   end
+  # end
+  #
+  # def total_goals
+  #   goal_count = 0
+  #   games.each do |game|
+  #     goal_count += game['home_goals'].to_i
+  #     goal_count += game['away_goals'].to_i
+  #   end
+  #   return goal_count
+  # end
+  #
+  # def total_number_of_games
+  #   game_count = 0
+  #   games.each do |game|
+  #     game_count += 1
+  #   end
+  #   return game_count
+  # end
 
-  def total_number_of_goals
-    home_goals.count + away_goals.count
-  end
-
-  def home_goals
-    games.find_all do |game|
-      game['away_goals'].to_i
-    end
-  end
-
-  def total_goals
-    goal_count = 0
-    games.each do |game|
-      goal_count += game['home_goals'].to_i
-      goal_count += game['away_goals'].to_i
-    end
-    return goal_count
-  end
-
-  def total_number_of_games
-    game_count = 0
-    games.each do |game|
-      game_count += 1
-    end
-    return game_count
-  end
-
-  def average_goals_by_season
-
-    season_hash = get_total_goals
-    season_hash.each do |season, goals|
-      count_of_games_by_season.each do |season_games, games|
-        if season_games == season
-          average = goals.to_f / games.to_f
-          season_hash[season] = (average).round(2)
-        end
-      end
-    end
-  end
-
-  def get_total_goals
-    season_hash = create_season_hash
-    season_hash_with_goals = season_hash.each do |season, total_goals|
-      require "pry"; binding.pry
-      games.each do |game|
-        if game['season'] == season
-          total_goals += game['away_goals'].to_i
-          total_goals += game['home_goals'].to_i
-        end
-      end
-      season_hash_with_goals[season] = total_goals
-    end
-  end
-
-  def create_season_hash
-    season_hash = games.each do |game|
-      season_hash[game['season']] = 0
-    end
-    season_hash
-  end
+  # def average_goals_by_season
+  #
+  #   season_hash = get_total_goals
+  #   season_hash.each do |season, goals|
+  #     count_of_games_by_season.each do |season_games, games|
+  #       if season_games == season
+  #         average = goals.to_f / games.to_f
+  #         season_hash[season] = (average).round(2)
+  #       end
+  #     end
+  #   end
+  # end
+  #
+  # def get_total_goals
+  #   season_hash = create_season_hash
+  #   season_hash_with_goals = season_hash.each do |season, total_goals|
+  #     require "pry"; binding.pry
+  #     games.each do |game|
+  #       if game['season'] == season
+  #         total_goals += game['away_goals'].to_i
+  #         total_goals += game['home_goals'].to_i
+  #       end
+  #     end
+  #     season_hash_with_goals[season] = total_goals
+  #   end
+  # end
+  #
+  # def create_season_hash
+  #   season_hash = games.each do |game|
+  #     season_hash[game['season']] = 0
+  #   end
+  #   season_hash
+  # end
 
 
 #
